Refactor: Moved `helpers` to `core.lib` and `core.utils`
The `core.lib` module has also been refactored to define functions after
defining the `lib` table, which is required since some functions call
functions in the same module. This was doable before the refactor
because all functions were global objects, but since they're now local,
Lua cannot find a symbol in a table it has not yet finished creating.
