+++ lua/neorg/core/init.lua
@@ -1,5 +1,7 @@
 local neorg = {
     callbacks = require("neorg.core.callbacks"),
+    lib = require("neorg.core.lib"),
+    utils = require("neorg.core.utils"),
 }
 
 return neorg
