+++ lua/neorg/events.lua
@@ -9,7 +9,6 @@ local neorg = require("neorg.core")
 local log = require("neorg.external.log")
 
 require("neorg.modules")
-require("neorg.external.helpers")
 
 neorg.events = {}
